removed redux connect from all components
