+++ App.js
@@ -6,10 +6,14 @@ import { AppLoading, Asset, Font } from 'expo';
 import { Ionicons } from '@expo/vector-icons';
 // import RootNavigation from './navigation/RootNavigation';
 import Login from './components/Login/Login'
+import { connect } from 'react-redux'
+
+
 
 export default class App extends React.Component {
   state = {
     isLoadingComplete: false,
+
   };
 
   render() {
@@ -65,3 +69,7 @@ const styles = StyleSheet.create({
     backgroundColor: '#fff',
   },
 });
+
+
+
+// export default connect()(App)
