+++ package.json
@@ -13,11 +13,15 @@
   },
   "dependencies": {
     "@expo/samples": "2.1.1",
+    "bcrypt": "^2.0.1",
     "expo": "^27.0.1",
+    "jsonwebtoken": "^8.2.1",
+    "jwt-decode": "^2.2.0",
     "react": "16.3.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-27.0.0.tar.gz",
     "react-navigation": "2.0.0",
     "react-redux": "^5.0.7",
+    "redux": "^4.0.0",
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
