+++ redux/reducers/index.js
@@ -1,41 +1,7 @@
-const initialState = {
-    isFetching: false,
-    isAuthenticated: isAuthenticated(),
-    user: getUserTokenInfo(),
-    errorMessage: ''
-  }
-  
-  export default function auth (state = initialState, action) {
-    switch (action.type) {
-      case 'LOGIN_REQUEST':
-        return {
-          ...state,
-          isFetching: true,
-          isAuthenticated: false,
-          errorMessage: ''
-        }
-      case 'LOGIN_SUCCESS':
-        return {
-          ...state,
-          isFetching: false,
-          isAuthenticated: true,
-          user: action.user
-        }
-      case 'LOGIN_FAILURE':
-        return {
-          ...state,
-          isFetching: false,
-          isAuthenticated: false,
-          errorMessage: action.message
-        }
-      case 'LOGOUT_SUCCESS':
-        return {
-          ...state,
-          isFetching: false,
-          isAuthenticated: false,
-          user: null
-        }
-        default:
-      return state
-  }
-}
+import { combineReducers } from 'redux'
+
+import auth from './auth'
+
+export default combineReducers({
+  auth
+})
